dwh/src/requirements.txt
pyspark
python-dotenv
wheel
schedule
black
psycopg2-binary
google-cloud-bigquery